backend/api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["ny-bin <ny.bin1210@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.76.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
black = "^22.3.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
SQLAlchemy = "^1.4.36"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-decouple = "^3.6"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
target-version = ['py39']
line-length = 120

[tool.isort]
line_length = 120
multi_line_output = 3
include_trailing_comma = true
known_local_folder=['config',]

[tool.flake8]
max-line-length = 120
max-complexity = 18
ignore = "E203,E266,W503,"
